renamed adjustTestTime -> adjustCurrentTestTimeout
