+++ packages/testing/src/test/mocha.unit.ts
@@ -1,13 +1,12 @@
 import { expect } from 'chai';
 import { mochaCtx } from '../mocha-ctx';
-import { overrideDebugMode } from '../debug-tests';
-import { adjustTestTime } from '../timeouts';
+import { adjustCurrentTestTimeout, overrideDebugMode } from '../timeouts';
 
 describe('adjustTestTime', () => {
     it('increases the running test timeout', function () {
         overrideDebugMode(false);
         const originalTimeout = this.timeout();
-        expect(adjustTestTime(100)).to.equal(100);
+        expect(adjustCurrentTestTimeout(100)).to.equal(100);
         expect(this.timeout()).to.equal(originalTimeout + 100);
     });
 });
